Add keymaps to allow toggling lines on both sides
Currently if you want to toggle line-by-line you need to navigate to the
'other' side of the diff to toggle the corresponding line there.

The more natural experience would be to have a keymap which toggles the
corresponding line (from the perspective of the current hunk) on the
other side of the diff at the same time.

This patch adds the keymap `toggle_line_pair` which is default mapped to
`s` in normal and visual mode which exposes this behaviour.